pyproject.toml
[project]
name = "resume-customizer"
version = "0.1.0"
description = "AI-powered resume customizer using Google Gemini API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.0.0",
    "python-docx>=1.1.0",
    "PyPDF2>=3.0.0",
    "python-dotenv>=1.0.0",
    "pypandoc>=1.13",
]

[project.optional-dependencies]
dev = [
    "streamlit>=1.28.0",
]
